requirements.txt
streamlit
python-dotenv
google-generativeai
pypdf2

langchain  # core frameworks
langchain-community # extra integration for calling models
faiss-cpu # Fast vector databse to store the embedded data
langchain-huggingface # connect huggingface models to perform embeddings
langchain-text-splitters # to split the data into smaller chunks
sentence-transformers # pre-trained models to convert chunks into vectors
langchain-core # to handle documents, chains of data etc..,